+++ tests/unit/fixtures/nc-text-input.js
@@ -1,0 +1,56 @@
+export const testText = 'Test Text'
+
+export const defaultProps = {
+  inputType: 'text',
+  error: '',
+  extraText: '',
+  containerBorderColor: '$containerBorderColor',
+  containerIsFocusBorderColor: '$containerIsFocusColor',
+  inputContentIsFocusLabelColor: '$containerIsFocusColor',
+  inputContentLabelColor: '$inputContentLabelColor',
+  errorColor: '$errorColor',
+  hasIconLeft: false,
+  hasIconRight: false,
+  hasIconRightOnFocus: false,
+  value: undefined,
+  disabled: false,
+  required: false,
+  name: undefined,
+  maxLength: 15,
+  label: 'label',
+  uiReference: 'uiEl',
+  iconLeftHasPointer: false,
+  iconRightHasPointer: false,
+  inputOptions: {},
+  id: undefined,
+  inputClasses: undefined,
+  wrapperClasses: undefined
+}
+
+export const propsWithValue = {
+  inputType: 'text',
+  error: '',
+  extraText: '',
+  containerBorderColor: '$containerBorderColor',
+  containerIsFocusBorderColor: '$containerIsFocusColor',
+  inputContentIsFocusLabelColor: '$containerIsFocusColor',
+  inputContentLabelColor: '$inputContentLabelColor',
+  errorColor: '$errorColor',
+  hasIconLeft: false,
+  hasIconRight: false,
+  hasIconRightOnFocus: false,
+  value: testText,
+  disabled: false,
+  required: false,
+  name: undefined,
+  maxLength: 15,
+  label: 'label',
+  uiReference: 'uiEl',
+  iconLeftHasPointer: false,
+  iconRightHasPointer: false,
+  inputOptions: {},
+  id: undefined,
+  inputClasses: undefined,
+  wrapperClasses: undefined
+}
+